projects/gemini3-hackathon/requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-multipart
pydantic
python-dotenv